Updated libmorton to latest version (with optional BMI2 support)

+++ src/libs/libmorton/include/morton_LUT_generators.h
@@ -8,7 +8,7 @@ template <typename element>
 void printTable(const element* table, size_t howmany, unsigned int splitat){
 	for (size_t i = 0; i < howmany; i++){
 		if (i % splitat == 0){ cout << endl; }
-		printf("%i ,", table[i]);
+		printf("%u ,", static_cast<unsigned int>(table[i]));
 	}
 	cout << endl;
 }
@@ -73,7 +73,7 @@ void generate3D_EncodeLUT(size_t how_many_bits, uint_fast32_t*& x_table, uint_fa
 }
 
 // Generate a decode tables for 3D morton code
-// how_many_bits should be a multiple of three
+// how_many_bits should be a multiple of three, to ensure you can shift results together
 void generate3D_DecodeLUT(size_t how_many_bits, uint_fast8_t*& x_table, uint_fast8_t*& y_table, uint_fast8_t*& z_table, bool print_tables){
 	size_t total = 1 << how_many_bits;
 	x_table = (uint_fast8_t*) malloc(total * sizeof(uint_fast8_t));
